User-related classes terminated. Code rectification. Relations between tables. Exceptions. Package aspects.

+++ Hello-Mam/src/main/java/com/mindera/HelloMam/repositories/RatingRepository.java
@@ -4,6 +4,8 @@ import com.mindera.HelloMam.entities.Rating;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface RatingRepository extends JpaRepository<Rating, Integer> {
+public interface RatingRepository extends JpaRepository<Rating, Long> {
 }
